+++ src/frontend/servers/y-provider/__tests__/convert.test.ts
@@ -18,6 +18,9 @@ import {
   COLLABORATION_SERVER_ORIGIN as origin,
 } from '../src/env';
 
+const expectedMarkdown = '# Example document\n\nLorem ipsum dolor sit amet.';
+const expectedHTML =
+  '<h1>Example document</h1><p>Lorem ipsum dolor sit amet.</p>';
 const expectedBlocks = [
   {
     children: [],
@@ -121,23 +124,43 @@ describe('Server Tests', () => {
     });
   });
 
+  test('POST /api/convert with unsupported Content-Type returns 415', async () => {
+    const app = initApp();
+    const response = await request(app)
+      .post('/api/convert')
+      .set('origin', origin)
+      .set('authorization', apiKey)
+      .set('content-type', 'image/png')
+      .send('randomdata');
+    expect(response.status).toBe(415);
+    expect(response.body).toStrictEqual({ error: 'Unsupported Content-Type' });
+  });
+
+  test('POST /api/convert with unsupported Accept returns 406', async () => {
+    const app = initApp();
+    const response = await request(app)
+      .post('/api/convert')
+      .set('origin', origin)
+      .set('authorization', apiKey)
+      .set('content-type', 'text/markdown')
+      .set('accept', 'image/png')
+      .send('# Header');
+    expect(response.status).toBe(406);
+    expect(response.body).toStrictEqual({ error: 'Unsupported format' });
+  });
+
   test.each([[apiKey], [`Bearer ${apiKey}`]])(
     'POST /api/convert with correct content with Authorization: %s',
     async (authHeader) => {
       const app = initApp();
 
-      const document = [
-        '# Example document',
-        '',
-        'Lorem ipsum dolor sit amet.',
-        '',
-      ].join('\n');
-
       const response = await request(app)
         .post('/api/convert')
         .set('Origin', origin)
         .set('Authorization', authHeader)
-        .send(document);
+        .set('content-type', 'text/markdown')
+        .set('accept', 'application/vnd.yjs.doc')
+        .send(expectedMarkdown);
 
       expect(response.status).toBe(200);
       expect(response.body).toBeInstanceOf(Buffer);
@@ -150,4 +173,95 @@ describe('Server Tests', () => {
       expect(blocks).toStrictEqual(expectedBlocks);
     },
   );
+
+  test('POST /api/convert Yjs to HTML', async () => {
+    const app = initApp();
+    const editor = ServerBlockNoteEditor.create();
+    const blocks = await editor.tryParseMarkdownToBlocks(expectedMarkdown);
+    const yDocument = editor.blocksToYDoc(blocks, 'document-store');
+    const yjsUpdate = Y.encodeStateAsUpdate(yDocument);
+    const response = await request(app)
+      .post('/api/convert')
+      .set('origin', origin)
+      .set('authorization', apiKey)
+      .set('content-type', 'application/vnd.yjs.doc')
+      .set('accept', 'text/html')
+      .send(Buffer.from(yjsUpdate));
+    expect(response.status).toBe(200);
+    expect(response.header['content-type']).toBe('text/html; charset=utf-8');
+    expect(typeof response.text).toBe('string');
+    expect(response.text).toBe(expectedHTML);
+  });
+
+  test('POST /api/convert Yjs to Markdown', async () => {
+    const app = initApp();
+    const editor = ServerBlockNoteEditor.create();
+    const blocks = await editor.tryParseMarkdownToBlocks(expectedMarkdown);
+    const yDocument = editor.blocksToYDoc(blocks, 'document-store');
+    const yjsUpdate = Y.encodeStateAsUpdate(yDocument);
+    const response = await request(app)
+      .post('/api/convert')
+      .set('origin', origin)
+      .set('authorization', apiKey)
+      .set('content-type', 'application/vnd.yjs.doc')
+      .set('accept', 'text/markdown')
+      .send(Buffer.from(yjsUpdate));
+    expect(response.status).toBe(200);
+    expect(response.header['content-type']).toBe(
+      'text/markdown; charset=utf-8',
+    );
+    expect(typeof response.text).toBe('string');
+    expect(response.text.trim()).toBe(expectedMarkdown);
+  });
+
+  test('POST /api/convert Yjs to JSON', async () => {
+    const app = initApp();
+    const editor = ServerBlockNoteEditor.create();
+    const blocks = await editor.tryParseMarkdownToBlocks(expectedMarkdown);
+    const yDocument = editor.blocksToYDoc(blocks, 'document-store');
+    const yjsUpdate = Y.encodeStateAsUpdate(yDocument);
+    const response = await request(app)
+      .post('/api/convert')
+      .set('origin', origin)
+      .set('authorization', apiKey)
+      .set('content-type', 'application/vnd.yjs.doc')
+      .set('accept', 'application/json')
+      .send(Buffer.from(yjsUpdate));
+    expect(response.status).toBe(200);
+    expect(response.header['content-type']).toBe(
+      'application/json; charset=utf-8',
+    );
+    expect(Array.isArray(response.body)).toBe(true);
+    expect(response.body).toStrictEqual(expectedBlocks);
+  });
+
+  test('POST /api/convert Markdown to JSON', async () => {
+    const app = initApp();
+    const response = await request(app)
+      .post('/api/convert')
+      .set('origin', origin)
+      .set('authorization', apiKey)
+      .set('content-type', 'text/markdown')
+      .set('accept', 'application/json')
+      .send(expectedMarkdown);
+    expect(response.status).toBe(200);
+    expect(response.header['content-type']).toBe(
+      'application/json; charset=utf-8',
+    );
+    expect(Array.isArray(response.body)).toBe(true);
+    expect(response.body).toStrictEqual(expectedBlocks);
+  });
+
+  test('POST /api/convert with invalid Yjs content returns 400', async () => {
+    const app = initApp();
+    const response = await request(app)
+      .post('/api/convert')
+      .set('origin', origin)
+      .set('authorization', apiKey)
+      .set('content-type', 'application/vnd.yjs.doc')
+      .set('accept', 'application/json')
+      .send(Buffer.from('notvalidyjs'));
+    expect(response.status).toBe(400);
+    expect(response.body).toStrictEqual({ error: 'Invalid Yjs content' });
+  });
 });